+++ Compilers/yanc_version.h
@@ -8,6 +8,6 @@
 #ifndef YANC_VERSION_H
 #define YANC_VERSION_H
 
-#define YANC_VERSION "4.4"
+#define YANC_VERSION "4.4.1"
 
 #endif